+++ com.unity.render-pipelines.high-definition/Editor/Lighting/HDLightUI.Skin.cs
@@ -85,7 +85,8 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             public readonly GUIContent shadowTint = new GUIContent("Tint", "Tint the shadows. This option affect dynamic shadows, contact shadows and shadow mask. It don't affect baked shadows.");
             public readonly GUIContent contactShadows = new GUIContent("Enable", "Enable support for Contact Shadows on this Light. This is better for lights with a lot of visible shadows.");
             public readonly GUIContent shadowUpdateMode = new GUIContent("Update Mode", "Sets the frequency of update for the shadow map.");
-
+            public readonly GUIContent useCustomSpotLightShadowCone = new GUIContent("Custom Spot Angle", "If selected, the custom spot angle will be used for shadow map rendering.");
+            public readonly GUIContent customSpotLightShadowCone = new GUIContent("Shadow Angle", "Custom angle used for shadow map rendering.");
 
             // Bias control
             public readonly GUIContent viewBiasMin = new GUIContent("View Bias");